requirements.txt
streamlit==1.46.0
requests==2.32.4
langchain==0.3.26
langchain-core==0.3.74
langchain-community==0.3.20
langgraph==0.6.5
langgraph-checkpoint==2.1.0
langgraph-checkpoint-sqlite==2.0.10
langchain-mcp-adapters
langchain-groq==0.3.2
langchain-tavily==0.2.11
langsmith==0.3.45
fastapi==0.116.0
uvicorn==0.35.0
geocoder==1.38.1
geopy==2.4.1
python-dotenv==1.1.0
groq
streamlit-audiorec
wordcloud
weave
wandb